no subdirectory
wait and see if we need that kind of organization

(also gets rid of require_relative since was introduced in Ruby 1.9)
