+++ lib/lolcommits.rb
@@ -15,8 +15,7 @@ require 'lolcommits/capturer'
 require 'lolcommits/git_info'
 require 'lolcommits/installation'
 require 'lolcommits/plugin'
-
-require 'lolcommits/helpers/platform'
+require 'lolcommits/platform'
 
 require 'lolcommits/plugins/loltext'
 require 'lolcommits/plugins/dot_com'
